requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
Faker==37.1.0
inflection==0.5.1
Markdown==3.8
packaging==25.0
pillow==11.2.1
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
uritemplate==4.1.1
